feat: subindo git pages

+++ src/app/detail/detail.service.ts
@@ -12,26 +12,58 @@ export class DetailService {
   ) { }
 
   getPost(id: any) {
-    return this.http.get(environment.urlApi + '/' + environment.nameBlog + '/' + id + '.json', {
-      headers: new HttpHeaders({
-        'Content-Type': 'application/json',
-      }),
-    }).toPromise();
+    if (environment.production) {
+      return this.http.get(environment.urlApi + '/' + environment.nameBlog + '/' + id + '.json', {
+        headers: new HttpHeaders({
+          'Content-Type': 'application/json',
+        }),
+      }).toPromise();
+    }
+    return new Promise((resolve) => {
+      setTimeout(() => {
+        const data = sessionStorage.getItem(environment.nameBlog);
+        const posts = data ? JSON.parse(data) : [];
+        const post = posts[id];
+        resolve(post);
+      }, Math.random() * 1000);
+    });
   }
 
   savePost(id: any, post: any) {
-    return this.http.put(environment.urlApi + '/' + environment.nameBlog + '/' + id + '.json', post, {
-      headers: new HttpHeaders({
-        'Content-Type': 'application/json',
-      }),
-    }).toPromise();
+    if (environment.production) {
+      return this.http.put(environment.urlApi + '/' + environment.nameBlog + '/' + id + '.json', post, {
+        headers: new HttpHeaders({
+          'Content-Type': 'application/json',
+        }),
+      }).toPromise();
+    }
+    return new Promise((resolve) => {
+      setTimeout(() => {
+        const data = sessionStorage.getItem(environment.nameBlog);
+        const posts = data ? JSON.parse(data) : [];
+        posts[id] = post;
+        sessionStorage.setItem(environment.nameBlog, JSON.stringify(posts));
+        resolve(post);
+      }, Math.random() * 1000);
+    });
   }
 
   deletePost(id: any) {
-    return this.http.delete(environment.urlApi + '/' + environment.nameBlog + '/' + id + '.json', {
-      headers: new HttpHeaders({
-        'Content-Type': 'application/json',
-      }),
-    }).toPromise();
+    if (environment.production) {
+      return this.http.delete(environment.urlApi + '/' + environment.nameBlog + '/' + id + '.json', {
+        headers: new HttpHeaders({
+          'Content-Type': 'application/json',
+        }),
+      }).toPromise();
+    }
+    return new Promise((resolve) => {
+      setTimeout(() => {
+        const data = sessionStorage.getItem(environment.nameBlog);
+        const posts = data ? JSON.parse(data) : [];
+        delete posts[id];
+        sessionStorage.setItem(environment.nameBlog, JSON.stringify(posts));
+        resolve([]);
+      }, Math.random() * 1000);
+    });
   }
 }
